requirements.txt
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.3
feedparser==6.0.2
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.6.3
mccabe==0.6.1
PlexAPI==4.5.2
pycodestyle==2.6.0
pylint==2.5.3
python-qbittorrent==0.4.2
requests==2.25.1
sgmllib3k==1.0.0
six==1.15.0
toml==0.10.1
urllib3==1.26.4
wincertstore==0.2
wrapt==1.12.1
